requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
bcrypt<4.0.0
pytest
httpx
requests
aiofiles
python-multipart
PyJWT